Added wb_xcache_wrapper and wb_interco.
The wb_xcache_wrapper is a iss2 wrapper the to a WishBone interface. The code
is similar to the vci_xcache_wrapper exept that it generates wb transactions.

The wb_interco is a simple WB interconnect, with a similar behaviour to the
WISHBONE Conbus IP Core that we can find on OpenCores web site:
"http://www.opencores.org/project,wb_conbus" except that we are limited to 32
masters/slaves instead of 8.
    
    The arbiter is quite simple,  priority is round robin starting from the
    first master (#0) and assumes that masters are faire.

Added also caba-wb-hello_world-iss2_lm32, a platform to demonstrate who to use
these components with the rest of SoCLib infrastructure.

The platform uses:

    One wb_xcache_wrapper around a big endian LM32 One wb_interco 2 VCI RAM
    1 VCI TTY 3 WB->VCI wrappers

The soft and behaviour is similar to the caba-vci-hello_world-iss2_lm32 which
uses a little endian LM32.
    


git-svn-id: https://www.soclib.fr/svn/trunk/soclib@1707 2cf4c1f0-a32b-0410-9068-d80f8be93c97