Introduce read_neutral(), which is a read() without LRU update. To be used in diagnostic/consistency checks where the cache state should not be changed.
victim_select(): do not invalidate the cache entry, it has to be
done on the next cycle (existing caches have been updated to deal with
this). This fixes a bug in the vci_ccvache_wrapper_v4, which would
consult the cache on next cycle TLB entries have to be invalidated.


git-svn-id: https://www.soclib.fr/svn/trunk/soclib@2169 2cf4c1f0-a32b-0410-9068-d80f8be93c97