Introducing write_pipe_frozen signal which is assigned in the P1 stage of pipeline to inform the P0 stage that the pending write in the P1 stage have not be treated because it is not possible to write in the write buffer or because there is a pending uncacheable write.
The signal write_pipe_frozen is used in the P0 stage to know if we can accept a new
request from the processor.

Introducing also in several places of the DCACHE_IDLE state, the verification of the
fillness state of the write pipe-line. As a general rule, before abandon the
DCACHE_IDLE state, we must verify that the pipe-line is empty, otherwise, we must
wait until it is the case.

Introducing in the condition to treat a uncacheable read in the VCI_CMD_FSM,
the verification of a miss in the write buffer for the same address.
Considering that the processor is not frozen after a uncacheable write, if after this
write, the processor request a uncacheable read on the same address, we must verify
that the write is sent before the read.



git-svn-id: https://www.soclib.fr/svn/trunk/soclib@2270 2cf4c1f0-a32b-0410-9068-d80f8be93c97