Improve the code replication mechanism on the vci_simple_rom component
This mechanism allows to drop some of most significant bits of the
allocated segments (defined by the nb_msb_drop parameter) to allow a ROM
to manage several segments with the same offset.

This mecanism can be used to replicate a code on several ROMs for a
clusterized architecture. If all the ROMs are on the same offset in all
clusters and the global bits (cluster coordinates) of the address are
masked, the code will be replicated.

Another use of this mechanism is for example fault tolerance. In a
clusterized architecture, if a cluster is faulty, the commands to the
faulty cluster can be redirected to another cluster when the target
component is replicated.

git-svn-id: https://www.soclib.fr/svn/trunk/soclib@2620 2cf4c1f0-a32b-0410-9068-d80f8be93c97